Move BaseTableItem into its own file
